# apps/backend/src/controller/invoice.controller.ts
import {
  Controller,
  Get,
  Post,
  Put,
  Del,
  Body,
  Param,
  Query,
  Inject,
} from '@midwayjs/decorator';
import { Context } from '@midwayjs/koa';
import { Validate } from '@midwayjs/validate';
import { InvoiceService } from '../service/invoice.service';
import {
  CreateInvoiceDto,
  UpdateInvoiceDto,
  VoidInvoiceDto,
  PaginationQuery,
  ApiResponse,
} from '../interface';

@Controller('/api/v1/invoices')
export class InvoiceController {
  @Inject()
  invoiceService: InvoiceService;

  @Inject()
  ctx: Context;

  /**
   * 创建发票
   */
  @Post('/')
  @Validate()
  async createInvoice(
    @Body() createInvoiceDto: CreateInvoiceDto
  ): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;

      if (!kitId) {
        return {
          success: false,
          message: '请选择套装或选择有效的合同',
          code: 400,
        };
      }

      const invoice = await this.invoiceService.createInvoice(
        createInvoiceDto,
        kitId
      );
      return {
        success: true,
        data: invoice,
        message: '发票创建成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '发票创建失败',
        code: 400,
      };
    }
  }

  /**
   * 获取发票列表
   */
  @Get('/')
  async getInvoices(
    @Query()
    query: PaginationQuery & {
      contractId?: number;
      status?: string;
      viewAll?: string;
    }
  ): Promise<ApiResponse> {
    try {
      // 如果 viewAll=true，则不传 kitId（查看全部套账）
      const kitId =
        query.viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const result = await this.invoiceService.getInvoices(query, kitId);
      return {
        success: true,
        data: result,
        message: '获取发票列表成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取发票列表失败',
        code: 500,
      };
    }
  }

  /**
   * 根据ID获取发票详情
   */
  @Get('/:id')
  async getInvoiceById(
    @Param('id') id: number,
    @Query('viewAll') viewAll?: string
  ): Promise<ApiResponse> {
    try {
      const kitId = viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const invoice = await this.invoiceService.getInvoiceById(id, kitId);
      if (!invoice) {
        return {
          success: false,
          message: '发票不存在',
          code: 404,
        };
      }
      return {
        success: true,
        data: invoice,
        message: '获取发票详情成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取发票详情失败',
        code: 500,
      };
    }
  }

  /**
   * 更新发票信息
   */
  @Put('/:id')
  @Validate()
  async updateInvoice(
    @Param('id') id: number,
    @Body() updateInvoiceDto: UpdateInvoiceDto
  ): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      if (!kitId || !this.ctx.state?.user?.id)
        throw new Error('请登录并选择当前套账');
      const invoice = await this.invoiceService.updateInvoice(
        id,
        updateInvoiceDto,
        kitId
      );
      if (!invoice) {
        return {
          success: false,
          message: '发票不存在',
          code: 404,
        };
      }
      return {
        success: true,
        data: invoice,
        message: '发票信息更新成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '发票信息更新失败',
        code: 400,
      };
    }
  }

  /**
   * 删除发票
   */
  @Del('/:id')
  async deleteInvoice(@Param('id') id: number): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      if (!kitId || !this.ctx.state?.user?.id)
        throw new Error('请登录并选择当前套账');
      const success = await this.invoiceService.deleteInvoice(id, kitId);
      if (!success) {
        return {
          success: false,
          message: '发票不存在',
          code: 404,
        };
      }
      return {
        success: true,
        message: '发票删除成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '发票删除失败',
        code: 400,
      };
    }
  }

  @Post('/:id/void')
  @Validate()
  async voidInvoice(
    @Param('id') id: number,
    @Body() dto: VoidInvoiceDto
  ): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      const userId = this.ctx.state?.user?.id;
      if (!kitId || !userId) throw new Error('请登录并选择当前套账');
      const invoice = await this.invoiceService.voidInvoice(
        id,
        dto?.reason,
        kitId,
        userId
      );
      if (!invoice) return { success: false, message: '发票不存在', code: 404 };
      return { success: true, data: invoice, message: '发票已在台账中作废' };
    } catch (error) {
      return {
        success: false,
        message: error.message || '发票作废失败',
        code: 400,
      };
    }
  }

  /**
   * 获取发票统计信息
   */
  @Get('/stats/overview')
  async getInvoiceStats(
    @Query() query: { viewAll?: string }
  ): Promise<ApiResponse> {
    try {
      // 如果 viewAll=true，则不传 kitId（查看全部套账）
      const kitId =
        query.viewAll === 'true' ? undefined : this.ctx.state?.kitId;
      const stats = await this.invoiceService.getInvoiceStats(undefined, kitId);
      return {
        success: true,
        data: stats,
        message: '获取发票统计成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取发票统计失败',
        code: 500,
      };
    }
  }

  /**
   * 获取逾期发票
   */
  @Get('/overdue/list')
  async getOverdueInvoices(): Promise<ApiResponse> {
    try {
      const kitId = this.ctx.state?.kitId;
      const invoices = await this.invoiceService.getOverdueInvoices(kitId);
      return {
        success: true,
        data: invoices,
        message: '获取逾期发票成功',
      };
    } catch (error) {
      return {
        success: false,
        message: error.message || '获取逾期发票失败',
        code: 500,
      };
    }
  }
}