# docs/api/backend_service/api_spec.yaml
openapi: 3.0.3
info:
  title: ProContractLedger API (合同台账管理系统 API)
  version: v1.0.0
  description: |-
    用于管理ProContractLedger（合同台账管理系统）的客户、合同、发票和付款的API。
    此API基于提供的数据库结构和核心SQL查询进行设计。
  contact:
    name: API技术支持
    email: support@example.com
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: 开发服务器
  - url: https://api.procontractledger.com/v1
    description: 生产服务器
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT认证令牌。在请求头中添加：Authorization: Bearer <token>"
  schemas:
...
    InvoiceBase:
      type: object
      description: 基础发票信息。对应 `invoices` 表。
      properties:
        invoice_id:
          type: integer
          format: int32
          description: 发票的唯一标识符。
          readOnly: true
        contract_id:
          type: integer
          format: int32
          description: 关联合同的标识符。
        invoice_info_id:
          type: integer
          format: int32
          description: 使用的客户发票信息的标识符。
        invoice_number:
          type: string
          maxLength: 50
          description: 发票的唯一编号。
        amount:
          type: string # 十进制数使用字符串
          format: decimal
          description: 发票金额。
        issue_date:
          type: string
          format: date # 根据表结构使用date-time
          description: 发票开具日期。
        due_date:
          type: string
          format: date # 根据表结构使用date-time
          description: 发票付款到期日。
        status:
          type: string
          maxLength: 20
          description: cancelled 表示不可恢复的台账作废，必须使用专用作废接口。
          enum: [draft, sent, paid, overdue, cancelled, bad_debt]
        notes:
          type: string
          format: text
          nullable: true
          description: 关于发票的附加备注。
        created_at:
          type: string
          format: date-time
          description: 发票创建时间戳。
          readOnly: true
        void_reason:
          type: string
          maxLength: 500
          nullable: true
          readOnly: true
          description: 首次台账作废原因，历史数据可能为空。
        voided_by:
          type: integer
          nullable: true
          readOnly: true
          description: 首次作废操作人 ID。
        voided_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
      required:
        - contract_id
        - invoice_info_id
        - invoice_number
        - amount
        - issue_date
        - due_date
        - status
    InvoiceCreateRequest:
      type: object
      description: 创建新发票的请求体。
      properties:
        contract_id:
          type: integer
          format: int32
          description: 关联合同的标识符。
          example: 1
        invoice_info_id:
          type: integer
          format: int32
          description: 使用的客户发票信息的标识符。
          example: 1
        invoice_number:
          type: string
          maxLength: 50
          description: 发票的唯一编号。
          example: "INV-2023-001"
        amount:
          type: string
          format: decimal
          description: 发票金额。
          example: "5000.00"
        issue_date:
          type: string
          format: date
          description: 发票开具日期。
          example: "2023-03-15"
        due_date:
          type: string
          format: date
          description: 发票付款到期日。
          example: "2023-04-15"
        status:
          type: string
          maxLength: 20
          description: 发票当前状态 (例如：待开票, 已开票, 已收款)。
          enum: ["待开票", "已开票", "已收款"]
          example: "已开票"
        notes:
          type: string
          format: text
          nullable: true
          description: 关于发票的附加备注。
          example: "第一期款项"
      required:
        - contract_id
        - invoice_info_id
        - invoice_number
        - amount
        - issue_date
        - due_date
        - status
    InvoiceUpdateRequest:
      type: object
      description: 更新发票信息的请求体。所有字段都是可选的，只提供需要更新的字段。
      properties:
        invoice_info_id:
          type: integer
          format: int32
          description: 使用的客户发票信息的标识符。
          example: 2
        invoice_number:
          type: string
          maxLength: 50
          description: 发票的唯一编号。
          example: "INV-2023-001-REV"
        amount:
          type: string
          format: decimal
          description: 发票金额。
          example: "5500.00"
        issue_date:
          type: string
          format: date
          description: 发票开具日期。
          example: "2023-03-20"
        due_date:
          type: string
          format: date
          description: 发票付款到期日。
          example: "2023-04-20"
        status:
          type: string
          maxLength: 20
          description: 发票当前状态 (例如：待开票, 已开票, 已收款)。
          enum: ["待开票", "已开票", "已收款"]
          example: "已收款"
        notes:
          type: string
          format: text
          nullable: true
          description: 关于发票的附加备注。
          example: "已完成第一期款项，并调整金额"
    InvoiceResponse:
      type: object
      description: 发票信息响应体。
      allOf:
        - $ref: '#/components/schemas/InvoiceBase'
      properties:
        invoice_id:
          readOnly: true
        created_at:
          readOnly: true
        contract_name:
          type: string
          description: 关联合同的名称。
          readOnly: true
        customer_name:
          type: string
          description: 关联客户的名称。
          readOnly: true
        company_name:
          type: string
          description: 关联开票公司的名称。
          readOnly: true
        tax_number:
          type: string
          description: 关联开票公司的税号。
          readOnly: true
        bank_name:
          type: string
          description: 关联开票公司的开户银行名称。
          readOnly: true
        bank_account:
          type: string
          description: 关联开票公司的银行账号。
          readOnly: true
        address:
          type: string
          description: 关联开票公司的地址。
          readOnly: true
        phone:
          type: string
          description: 关联开票公司的电话。
          readOnly: true
        payments:
          type: array
          description: 关联的付款记录列表。
          items:
            type: object
            properties:
              payment_id:
                type: integer
                format: int32
              amount:
                type: string
                format: decimal
              payment_date:
                type: string
                format: date
              payment_method:
                type: string
              reference_number:
                type: string
                nullable: true
              notes:
                type: string
                nullable: true
              created_at:
                type: string
                format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /invoices:
    get:
      summary: 获取所有发票
      tags:
        - Invoices
      parameters:
        - in: query
          name: contract_id
          schema:
            type: integer
            format: int32
          description: 合同ID
        - in: query
          name: invoice_number
          schema:
            type: string
          description: 发票编号模糊搜索
        - in: query
          name: status
          schema:
            type: string
            enum: ["待开票", "已开票", "已收款"]
          description: 发票状态
        - in: query
          name: issue_date
          schema:
            type: string
            format: date
          description: 发票开具日期 (YYYY-MM-DD)
        - in: query
          name: due_date
          schema:
            type: string
            format: date
          description: 发票付款到期日 (YYYY-MM-DD)
        - in: query
          name: page
          schema:
            type: integer
            format: int32
            minimum: 1
          description: 页码，默认为1
        - in: query
          name: pageSize
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
          description: 每页数量，默认为10
      responses:
        200:
          description: 成功获取发票列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/InvoiceResponse'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
        500:
          description: 获取发票列表失败
    post:
      summary: 创建发票
      tags:
        - Invoices
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceCreateRequest'
      responses:
        201:
          description: 发票创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "发票创建成功"
                  invoice_id:
                    type: integer
                    format: int32
        400:
          description: 缺少必要的发票信息
        404:
          description: 关联的合同不存在 / 关联的发票信息不存在
        409:
          description: 发票号码已存在
        500:
          description: 创建发票失败
  /invoices/{invoiceId}/void:
    post:
      summary: 作废发票台账记录
      description: 不执行税务作废或红冲。当前授权套账内操作，不能用 viewAll 放宽写权限。原因去除首尾空格后为 1–500 字；已支付、存在待处理或已完成收款、坏账状态或坏账金额时拒绝。重复合法请求保留首次审计。
      tags: [Invoices]
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: integer
        - in: header
          name: X-Kit-Id
          schema:
            type: integer
          description: 当前套账，由中间件校验访问权限。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        200:
          description: 返回 ApiResponse；success 为 false 时检查 message 和业务 code，400 表示条件不满足，404 表示当前套账内不存在。
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/InvoiceBase'
                  message:
                    type: string
                  code:
                    type: integer
        401:
          description: 未登录。
        403:
          description: 无套账访问权限。
  /invoices/{invoiceId}:
    get:
      summary: 获取单个发票
      tags:
        - Invoices
      parameters:
        - in: path
          name: invoiceId
          schema:
            type: integer
            format: int32
          required: true
          description: 发票ID
      responses:
        200:
          description: 成功获取发票信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        404:
          description: 发票不存在
        500:
          description: 获取发票详情失败
    put:
      summary: 更新发票
      tags:
        - Invoices
      parameters:
        - in: path
          name: invoiceId
          schema:
            type: integer
            format: int32
          required: true
          description: 发票ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceUpdateRequest'
      responses:
        200:
          description: 发票更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "发票更新成功"
        400:
          description: 没有要更新的字段
        404:
          description: 发票不存在
        409:
          description: 发票号码已被其他发票使用
        500:
          description: 更新发票失败
    delete:
      summary: 删除发票
      tags:
        - Invoices
      parameters:
        - in: path
          name: invoiceId
          schema:
            type: integer
            format: int32
          required: true
          description: 发票ID
      responses:
        200:
          description: 发票删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "发票删除成功"
        404:
          description: 发票不存在
        409:
          description: 无法删除发票，存在关联的付款记录
        500:
          description: 删除发票失败
...